algorithms/requirements.txt
Flask==1.1.2
tensorflow==2.4.0
tensorflow-datasets==4.2.0
pillow==8.1.2
pandas==1.1.5
opencv-python==4.1.1.26
flask-ngrok==0.0.25
sk-video==1.1.10
scipy==1.1.0
streamlink==2.1.2
tf-slim==1.1.0
urllib3==1.26.4
pika==1.2.0